+++ CHANGELOG.md
@@ -47,6 +47,8 @@ The format is based on [Keep a Changelog] and this project adheres to
 
 - Core Foundation constants: `kCFNotFound`.
 
+- Methods for `NSStringEncoding`: `name`.
+
 - Foundation error codes.
 
 - `core_graphics` module for [Core Graphics](https://developer.apple.com/documentation/coregraphics)